requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

accelerate=0.28.0
cudatoolkit=11.8.0
datasets=2.12.0=pypi_0
diffusers=0.31.0.dev0
en-core-web-sm=3.7.1
flash-attn=2.5.8
gensim=4.3.2
huggingface-hub=0.24.7
jieba=0.42.1
nltk=3.8.1
numpy=1.23.5
pandas=1.5.3
peft=0.7.0
python=3.10.13
pytorch=1.13.1
scikit-learn=1.2.2
scipy=1.10.1
spacy=3.7.6
transformers=4.25.1
zh-core-web-sm=3.7.0
